Updated Login component to SignInPage for consistency. Added temporary buttons for dev ergonomics; these will go away soon. Some simplifications to CreateGigPanel.tsx as well. Also created and applied routes.ts file for sanity.
