+++ src/common/routes.ts
@@ -1,0 +1,4 @@
+export const routes = {
+    MAIN_PAGE: '/',
+    SIGN_IN_PAGE: 'sign-in',
+};
